# openapi.yaml
openapi: 3.0.3
info:
  title: Oblixa route contract
  version: 0.1.0
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: sb-access-token
    CronSecret:
      type: http
      scheme: bearer
    WebhookSignature:
      type: apiKey
      in: header
      name: stripe-signature
paths:
...
  /api/contact:
    post:
      operationId: post_contact
      summary: POST /api/contact
      description: Generated route contract for src/app/api/contact/route.ts.
      security:
        - SessionCookie: []
      responses:
        "200":
          description: OK
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
        "409":
          description: Conflict
        "413":
          description: Payload too large
        "429":
          description: Rate limited
        "500":
          description: Unexpected server error
...